am335x: Enable NOR device support
This patch enables the NOR support and configures
the environment settings for NOR as well as pin muxing and
startup configuration for XIP booting.

In NOR boot mode, RBL does pin muxing only for address lines
a0-a11. So, the pin muxing for remaining address lines have to
be done by the software before accessing the higher memory.

Signed-off-by: Sugumar Natarajan <sugumar@ti.com>