SPL: align TI81xx/AM335x with the OMAP4 romcode boot device detection
This uses the same detection method that OMAP4 uses to determine what
device the SPL image was loaded from. Although the scheme used is
the same, the actual values indicating load from MMC1, NAND, etc.
are different on this class of parts. Only MMC1 and NAND are verified
at this point.

Signed-off-by: Matt Porter <mporter@ti.com>