am335x: Clean up the config file, add uEnv.txt
This removes a bunch of legacy options that SPL now takes care of for
us, does some style clean-ups and adds uEnv.txt support based on
Beagleboard, along with optargs also from Beagleboard and corrects
the clock frequency for the timer.  To achieve faster boot times
we add kloadaddr which is the default kernel address minus the uImage
header and make use of this in the various possible bootcmds.  This
lets us avoid a copy of the kernel image.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>
Signed-off-by: Jason Kridner <jkridner@beagleboard.org>
Signed-off-by: Koen Kooi <k-kooi@ti.com>
Signed-off-by: Tom Rini <trini@ti.com>