AM3517 EVM: Add am3517_evm_norflash and _norflash_boot targets
These build targets will make a U-Boot that knows about the NOR
flash that can be enabled on this board (see previous commits for
details) and a U-Boot that can run from NOR flash (u-boot.bin).

When we are building to run from NOR flash we makes changes to
the default environment in order to run the system fully from
NOR flash.  This results in a lot of 'noise' in terms of moving
NAND-specific config options to be in one place that we can
key off of.  We also modify the bootcmd to use the flash the
user has chosen (by building for NOR, or NAND).

Signed-off-by: Tom Rini <trini@ti.com>