CPSW: Add gigabit_en flag based on the kernel driver, use
Using the same logic as the kernel driver, add a gigabit_en flag
that can be used to say if gigabit is supported or not.  Default
to enabled, clear for Bone (for safety), IA (which needs general
work atm) and 1.0 EVMs.  In the CPSW driver, use this flag to set
or clear bit 7 if we're running in gigabit mode.

Signed-off-by: Tom Rini <trini@ti.com>