Azure: Re-organize world build again
We're now hitting the place where sometimes our world build job fails
due to lack of space. Try and resolve this again by moving some boards
around. We move the TI K3 platforms to be with powerpc(which moves both
arm and aarch64 platforms). We also split sun7i boards out of the sunxi
job and move them to the multi-vendor arm32 job.

Signed-off-by: Tom Rini <trini@konsulko.com>